refactor(mobile): ME-11 BigDecimal migration za Account/Payment/Transfer/Loan DTOs (C2 §255 paritet)
Migrirano sa Double na BigDecimal za sva novcana polja u Account, Payment,
Transfer i Loan DTO familijama (spec C2 §255). Sada paritet sa FE/BE koji
vec koriste BigDecimal za istu domenu.

Production izmene:
- Novi BigDecimalAdapter — Moshi nema nativni adapter za java.math.BigDecimal,
  pa je do sada CardDto/SavingsDto deserializacija sa BE-a IIRC pucala u
  runtime-u (bug iz Faze A). Adapter registrovan u NetworkModule.provideMoshi
  pre KotlinJsonAdapterFactory.
- AccountDto: balance/availableBalance/reservedAmount/dailyLimit/monthlyLimit/
  initialDeposit prebaceni na BigDecimal. Dodat effectiveReserved helper
  koji vraca BigDecimal preferirajuci reservedAmount field, fallback na
  (balance - availableBalance).
- PaymentDto + PaymentListDto: amount/fee BigDecimal.
- TransferDto: amount/convertedAmount/fee BigDecimal; rate ostaje Double (FX
  koeficijent, ne novac).
- LoanDto + LoanInstallmentDto + LoanApplicationDto + Response: amount/balance/
  monthlyInstallment/monthlyIncome BigDecimal; interestRate/effectiveRate
  ostaju Double (procenti).
- InterbankDto + InitiateInterbankPaymentDto: amount/convertedAmount/fee
  BigDecimal; rate ostaje Double.

Consumer migracije (svi ViewModel-i + Compose ekrani):
- HomeViewModel.totalRsdBalance + balanceTrend → BigDecimal (sa precizniom
  aritmetikom koja izbegava 0.1 + 0.2 = 0.30000000000000004 Double bug).
  ClientHomeContent prima BigDecimal za hero totalRsd, mapira u Double samo
  na granici MiniBarChart (UI-only Canvas).
- AccountsListScreen + AccountDetailsScreen: koristi effectiveReserved
  helper sa BigDecimal poredjenjem (> BigDecimal.ZERO).
- AccountDetailsViewModel.submitLimitChange + AccountRepository.updateLimits
  + LimitDialog: parametre prebaceni na BigDecimal? + parseBigDecimal.
- AccountRequestNewViewModel + CreateAccountForClientViewModel: koriste
  parseBigDecimal umesto parse() i postavljaju BigDecimal.ZERO default.
- NewPaymentViewModel + NewTransferViewModel + LoanApplyViewModel:
  parsedAmount: BigDecimal? sa <= BigDecimal.ZERO validacijom; OTP flow
  propagira BigDecimal kroz private startIntraBank/Interbank/submitWithOtp.
- LoansScreen + LoanDetailsScreen + AllLoansScreen + LoanRequestsScreen:
  ?: 0.0 zamenjeno sa ?: BigDecimal.ZERO. MoneyFormatter.format BigDecimal
  overload (vec postojao posle Faze A) auto-dispatch-uje.
- SavingsNewDepositScreen + RecurringOrdersScreen: raw interpolacija
  ${acc.availableBalance} zamenjena sa MoneyFormatter.format() da bi se
  izbegla potencijalna scientific notation za velike BigDecimal vrednosti.

Novi testovi (24 nova testa kroz 4 fajla):
- AccountDtoTest (9): JSON parsing preciznosti (14 cifara bez Double roundoff),
  effectiveReserved helper, BigDecimal aritmetika (0.1 + 0.2 = 0.3 egzaktno),
  Limits/Request/Response DTO-ovi.
- PaymentDtoTest (5): create/response serijalizacija, running balance
  aritmetika, fee nullable.
- TransferDtoTest (4): internal/fx request, response sa odvojenim
  BigDecimal (money) i Double (rate) parsing-om.
- LoanDtoTest (6): puna LoanDto parser, installment aritmetika, application
  request/response.

Verifikacija:
- compileDebugKotlin --rerun-tasks: BUILD SUCCESSFUL u 1m 16s
- lintDebug: BUILD SUCCESSFUL, 0 errors / 0 warnings (warningsAsErrors=true)
- testDebugUnitTest: 245/245 PASS (221 baseline + 24 nova) — bez regresija
- assembleDebug: BUILD SUCCESSFUL u 17s

Out of scope (ne migrirano u ovoj rundi):
- OrderDto/ActuaryDto/OtcDto/PortfolioDto/FundDto/TaxDto/MarginDto/OptionDto/
  ListingDto/DividendDto/PriceAlertDto/RecurringOrderDto/WatchlistDto/
  ProfitBankDto/FundStatisticsDto — sledeca runda ako se trazi puna
  ME-11 migracija svih domena (ovi DTO-ovi nisu u spec-u explicit zahtevu
  C2 §255 koji se odnosi na banking flow).
- ExchangeRepository.calculate(Double) — ostaje Double jer je FX UI estimate,
  nije autoritativna preracun.

+++ app/src/main/java/rs/raf/banka2/mobile/data/dto/interbank/InterbankDto.kt
@@ -1,17 +1,22 @@
 package rs.raf.banka2.mobile.data.dto.interbank
 
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 
 /**
  * Inicijalizacija inter-bank placanja. Backend prepoznaje koja banka je
  * primalac iz prefiksa broja racuna (prve 3 cifre = routing number) i
  * pokrece 2-Phase Commit transakciju.
+ *
+ * ME-11: novcana polja prebacena sa Double na BigDecimal (spec C2 §255).
+ * Polje `rate` (FX kurs) ostaje Double — to je decimal koeficijent koji
+ * ne predstavlja iznos pa precizija nije kritican zahtev.
  */
 @JsonClass(generateAdapter = true)
 data class InitiateInterbankPaymentDto(
     val fromAccountId: Long,
     val toAccountNumber: String,
-    val amount: Double,
+    val amount: BigDecimal,
     val recipientName: String,
     val paymentPurpose: String,
     val paymentCode: String = "289",
@@ -26,12 +31,12 @@ data class InterbankTransactionDto(
     val routingNumber: String? = null,
     val fromAccount: String? = null,
     val toAccount: String? = null,
-    val amount: Double? = null,
+    val amount: BigDecimal? = null,
     val currency: String? = null,
-    val convertedAmount: Double? = null,
+    val convertedAmount: BigDecimal? = null,
     val convertedCurrency: String? = null,
-    val rate: Double? = null,
-    val fee: Double? = null,
+    val rate: Double? = null,            // FX kurs — Double je adekvatan
+    val fee: BigDecimal? = null,
     val message: String? = null,
     val updatedAt: String? = null
 )
